Move QueryBox component in Sidebar folder
