+++ static/src/components/Sidebar/QueryBox.jsx
